refactor(controller/wavs): extract missing sample tracking
